LayoutTests/fast/css-grid-layout/grid-item-end-after-get-set.html

<!DOCTYPE html>
<html>
<head>
<script>
if (window.testRunner)
    testRunner.overridePreference("WebKitCSSGridLayoutEnabled", 1);
</script>
<style>
.gridItemWithPositiveInteger {
    -webkit-grid-end: 10;
    -webkit-grid-after: 15;
}
.gridItemWithNegativeInteger {
    -webkit-grid-end: -10;
    -webkit-grid-after: -15;
}
.gridItemWithAuto {
    -webkit-grid-end: auto;
    -webkit-grid-after: auto;
}
</style>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<!-- The first has no properties set on it. -->
<div id="gridElement"></div>
<div class="gridItemWithPositiveInteger" id="gridItemWithPositiveInteger"></div>
<div class="gridItemWithNegativeInteger" id="gridItemWithNegativeInteger"></div>
<div class="gridItemWithAuto" id="gridItemWithAutoElement"></div>
<script>
    description('Test that setting and getting grid-end and grid-after works as expected');

    debug("Test getting -webkit-grid-end and -webkit-grid-after set through CSS");
    var gridElement = document.getElementById("gridElement");
    shouldBe("getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-end')", "'auto'");
    shouldBe("getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-column')", "'auto'");
    shouldBe("getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-after')", "'auto'");
    shouldBe("getComputedStyle(gridElement, '').getPropertyValue('-webkit-grid-row')", "'auto'");

    var gridItemWithPositiveInteger = document.getElementById("gridItemWithPositiveInteger");
    shouldBe("getComputedStyle(gridItemWithPositiveInteger, '').getPropertyValue('-webkit-grid-end')", "'10'");
    shouldBe("getComputedStyle(gridItemWithPositiveInteger, '').getPropertyValue('-webkit-grid-column')", "'10'");
    shouldBe("getComputedStyle(gridItemWithPositiveInteger, '').getPropertyValue('-webkit-grid-after')", "'15'");
    shouldBe("getComputedStyle(gridItemWithPositiveInteger, '').getPropertyValue('-webkit-grid-row')", "'15'");

    var gridItemWithNegativeInteger = document.getElementById("gridItemWithNegativeInteger");
    shouldBe("getComputedStyle(gridItemWithNegativeInteger, '').getPropertyValue('-webkit-grid-end')", "'-10'");
    shouldBe("getComputedStyle(gridItemWithNegativeInteger, '').getPropertyValue('-webkit-grid-column')", "'-10'");
    shouldBe("getComputedStyle(gridItemWithNegativeInteger, '').getPropertyValue('-webkit-grid-after')", "'-15'");
    shouldBe("getComputedStyle(gridItemWithNegativeInteger, '').getPropertyValue('-webkit-grid-row')", "'-15'");

    var gridItemWithAutoElement = document.getElementById("gridItemWithAutoElement");
    shouldBe("getComputedStyle(gridItemWithAutoElement, '').getPropertyValue('-webkit-grid-end')", "'auto'");
    shouldBe("getComputedStyle(gridItemWithAutoElement, '').getPropertyValue('-webkit-grid-column')", "'auto'");
    shouldBe("getComputedStyle(gridItemWithAutoElement, '').getPropertyValue('-webkit-grid-after')", "'auto'");
    shouldBe("getComputedStyle(gridItemWithAutoElement, '').getPropertyValue('-webkit-grid-row')", "'auto'");

    debug("");
    debug("Test the initial value");
    var element = document.createElement("div");
    document.body.appendChild(element);
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-end')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-column')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-after')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-row')", "'auto'");

    debug("");
    debug("Test getting and setting grid-end and grid-after through JS");
    element.style.webkitGridEnd = "18";
    element.style.webkitGridAfter = "66";
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-end')", "'18'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-column')", "'18'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-after')", "'66'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-row')", "'66'");

    element = document.createElement("div");
    document.body.appendChild(element);
    element.style.webkitGridEnd = "-55";
    element.style.webkitGridAfter = "-40";
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-end')", "'-55'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-column')", "'-55'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-after')", "'-40'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-row')", "'-40'");

    element = document.createElement("div");
    document.body.appendChild(element);
    element.style.webkitGridEnd = "auto";
    element.style.webkitGridAfter = "auto";
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-end')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-column')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-after')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-row')", "'auto'");

    debug("");
    debug("Test setting grid-end and grid-after back to 'auto' through JS");
    element.style.webkitGridEnd = "18";
    element.style.webkitGridAfter = "66";
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-end')", "'18'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-column')", "'18'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-after')", "'66'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-row')", "'66'");
    element.style.webkitGridEnd = "auto";
    element.style.webkitGridAfter = "auto";
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-end')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-column')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-after')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('-webkit-grid-row')", "'auto'");
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
